requirements.txt
numpy>=1.24
gymnasium>=0.29
gymnasium-robotics>=1.4
mujoco>=3.0.0

stable-baselines3[extra]>=2.3.0
torch>=2.2
tensorboard>=2.16

matplotlib>=3.8
pillow>=10.0
requests>=2.31
python-dotenv>=1.0
tqdm>=4.66
